backend/requirements.txt
amqp==5.2.0
aniso8601==9.0.1
Babel==2.14.0
billiard==4.2.0
blinker==1.7.0
celery==5.3.6
chardet==5.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.5.0
email-validator==2.1.0.post1
Flask==3.0.1
Flask-BabelEx==0.9.4
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.6
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.5
MarkupSafe==2.1.4
passlib==1.7.4
pillow==10.2.0
prompt-toolkit==3.0.43
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
reportlab==4.1.0
six==1.16.0
speaklater==1.3
SQLAlchemy==2.0.25
typing_extensions==4.9.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
WTForms==3.1.2
